# Improve perf of hash one-shots on Win10