--- a/src/libraries/Common/src/Interop/Windows/BCrypt/BCryptAlgorithmCache.cs
+++ b/src/libraries/Common/src/Interop/Windows/BCrypt/BCryptAlgorithmCache.cs
@@ -17,7 +17,7 @@ internal partial class Interop
             /// <summary>
             ///     Return a SafeBCryptAlgorithmHandle of the desired algorithm and flags. This is a shared handle so do not dispose it!
             /// </summary>
-            public static SafeBCryptAlgorithmHandle GetCachedBCryptAlgorithmHandle(string hashAlgorithmId, BCryptOpenAlgorithmProviderFlags flags)
+            public static SafeBCryptAlgorithmHandle GetCachedBCryptAlgorithmHandle(string hashAlgorithmId, BCryptOpenAlgorithmProviderFlags flags, out int hashSizeInBytes)
             {
                 // There aren't that many hash algorithms around so rather than use a LowLevelDictionary and guard it with a lock,
                 // we'll use a simple list. To avoid locking, we'll recreate the entire list each time an entry is added and replace it atomically.
@@ -29,7 +29,10 @@ internal partial class Interop
                 foreach (Entry entry in cache)
                 {
                     if (entry.HashAlgorithmId == hashAlgorithmId && entry.Flags == flags)
+                    {
+                        hashSizeInBytes = entry.HashSizeInBytes;
                         return entry.Handle;
+                    }
                 }
 
                 SafeBCryptAlgorithmHandle safeBCryptAlgorithmHandle;
@@ -37,14 +40,15 @@ internal partial class Interop
                 if (ntStatus != NTSTATUS.STATUS_SUCCESS)
                     throw Interop.BCrypt.CreateCryptographicException(ntStatus);
 
-                Entry[] newCache = new Entry[cache.Length + 1];
+                Array.Resize(ref cache, cache.Length + 1);
                 Entry newEntry = new Entry(hashAlgorithmId, flags, safeBCryptAlgorithmHandle);
-                Array.Copy(cache, newCache, cache.Length);
-                newCache[newCache.Length - 1] = newEntry;
+                cache[^1] = new Entry(hashAlgorithmId, flags, safeBCryptAlgorithmHandle);
 
                 // Atomically overwrite the cache with our new cache. It's possible some other thread raced to add a new entry with us - if so, one of the new entries
                 // will be lost and the next request will have to allocate it again. That's considered acceptable collateral damage.
-                _cache = newCache;
+                _cache = cache;
+
+                hashSizeInBytes = newEntry.HashSizeInBytes;
                 return newEntry.Handle;
             }
 
@@ -52,17 +56,37 @@ internal partial class Interop
 
             private struct Entry
             {
-                public Entry(string hashAlgorithmId, BCryptOpenAlgorithmProviderFlags flags, SafeBCryptAlgorithmHandle handle)
+                public unsafe Entry(string hashAlgorithmId, BCryptOpenAlgorithmProviderFlags flags, SafeBCryptAlgorithmHandle handle)
                     : this()
                 {
                     HashAlgorithmId = hashAlgorithmId;
                     Flags = flags;
                     Handle = handle;
+
+                    int hashSize;
+                    NTSTATUS ntStatus = Interop.BCrypt.BCryptGetProperty(
+                        handle,
+                        Interop.BCrypt.BCryptPropertyStrings.BCRYPT_HASH_LENGTH,
+                        &hashSize,
+                        sizeof(int),
+                        out int cbHashSize,
+                        0);
+
+                    if (ntStatus != NTSTATUS.STATUS_SUCCESS)
+                    {
+                        throw Interop.BCrypt.CreateCryptographicException(ntStatus);
+                    }
+
+                    Debug.Assert(cbHashSize == sizeof(int));
+                    Debug.Assert(hashSize > 0);
+
+                    HashSizeInBytes = hashSize;
                 }
 
                 public string HashAlgorithmId { get; private set; }
                 public BCryptOpenAlgorithmProviderFlags Flags { get; private set; }
                 public SafeBCryptAlgorithmHandle Handle { get; private set; }
+                public int HashSizeInBytes { get; private set; }
             }
         }
     }