Uses bcrypt pseudo-handles (see https://docs.microsoft.com/windows/win32/seccng/cng-algorithm-pseudo-handles) with the `BCryptHash` routine to get a modest perf improvement in the one-shot hash algortihms when running on Windows 10. For downlevel Windows platforms, continues to use the existing "catch `EntryPointNotFound` and fall back to `BCryptHashData`" logic.

The Win10-specific routine also has a safety net such that if any of the invariants are violated (unknown algorithm or destination too short), it falls back to the downlevel code paths for additional error handling.

No change on other OSes.

|   Method | Toolchain | HashAlgorithm | SourceLengthInBytes |     Mean |   Error |  StdDev | Ratio |
|--------- |---------- |-------------- |-------------------- |---------:|--------:|--------:|------:|
| HashData |    master |          SHA1 |                   0 | 336.4 ns | 0.71 ns | 0.66 ns |  1.00 |
| HashData |     proto |          SHA1 |                   0 | 288.8 ns | 2.17 ns | 1.81 ns |  0.86 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |          SHA1 |                  64 | 420.8 ns | 2.22 ns | 1.85 ns |  1.00 |
| HashData |     proto |          SHA1 |                  64 | 374.6 ns | 1.35 ns | 1.27 ns |  0.89 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |          SHA1 |                 256 | 651.0 ns | 3.40 ns | 3.02 ns |  1.00 |
| HashData |     proto |          SHA1 |                 256 | 609.0 ns | 4.11 ns | 3.85 ns |  0.94 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA256 |                   0 | 274.7 ns | 1.33 ns | 1.24 ns |  1.00 |
| HashData |     proto |        SHA256 |                   0 | 226.2 ns | 1.13 ns | 0.94 ns |  0.82 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA256 |                  64 | 302.8 ns | 1.53 ns | 1.43 ns |  1.00 |
| HashData |     proto |        SHA256 |                  64 | 255.2 ns | 0.51 ns | 0.45 ns |  0.84 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA256 |                 256 | 390.4 ns | 1.10 ns | 0.92 ns |  1.00 |
| HashData |     proto |        SHA256 |                 256 | 348.5 ns | 1.90 ns | 1.78 ns |  0.89 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA512 |                   0 | 493.0 ns | 2.43 ns | 2.27 ns |  1.00 |
| HashData |     proto |        SHA512 |                   0 | 455.6 ns | 1.32 ns | 1.10 ns |  0.92 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA512 |                  64 | 498.4 ns | 1.02 ns | 0.91 ns |  1.00 |
| HashData |     proto |        SHA512 |                  64 | 468.5 ns | 2.48 ns | 2.07 ns |  0.94 |
|          |           |               |                     |          |         |         |       |
| HashData |    master |        SHA512 |                 256 | 947.0 ns | 4.13 ns | 3.86 ns |  1.00 |
| HashData |     proto |        SHA512 |                 256 | 907.9 ns | 4.28 ns | 4.01 ns |  0.96 |