@GrabYourPitchforks 2020-08-08
@vcsjones I addressed your second feedback item by querying the hash size right after we create the algorithm, then shoving that information into the algorithm handle cache. It means we only take the lookup hit once. For the one-shot algos, I think it's still appropriate to hard-code the expected digest length since we need to maintain the friendly-name -\> pseudo-alg handle mapping ourselves anyway.