@bartonjs 2020-08-13
```
    System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest.CustomConnectionFactory_SyncRequest_Fails [FAIL]
      Assert.Equal() Failure
      Expected: HostNotFound
      Actual:   Unknown
      Stack Trace:
        /_/src/libraries/System.Net.Http/tests/FunctionalTests/SocketsHttpHandlerTest.cs(165,0): at System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest.CustomConnectionFactory_SyncRequest_Fails()
        /_/src/libraries/System.Net.Http/tests/FunctionalTests/SocketsHttpHandlerTest.cs(165,0): at System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest.CustomConnectionFactory_SyncRequest_Fails()
        --- End of stack trace from previous location ---
    System.Net.Http.Functional.Tests.SyncHttpHandler_HttpClientHandler_Authentication_Test.Proxy_DomainJoinedProxyServerUsesKerberos_Success [SKIP]
      Condition(s) not met: "IsDomainJoinedServerAvailable"
    System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest_Http2.CustomConnectionFactory_SyncRequest_Fails [FAIL]
      Assert.Equal() Failure
      Expected: HostNotFound
      Actual:   Unknown
      Stack Trace:
        /_/src/libraries/System.Net.Http/tests/FunctionalTests/SocketsHttpHandlerTest.cs(165,0): at System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest.CustomConnectionFactory_SyncRequest_Fails()
        /_/src/libraries/System.Net.Http/tests/FunctionalTests/SocketsHttpHandlerTest.cs(165,0): at System.Net.Http.Functional.Tests.SocketsHttpHandler_ConnectionFactoryTest.CustomConnectionFactory_SyncRequest_Fails()
        --- End of stack trace from previous location ---
```

This same test (or a very similar looking one) has failed on at least 3 reruns for the last commit of the PR.  Looks like it only happens on Win7, and I haven't seen it on other PRs (but maybe I'm unobservant).

I don't see how it could necessarily be related, but the consistency bothers me.