@bartonjs 2020-08-13
Odd, I thought that every test rerun tried stacking the PR diff on top of master freshly... but I agree with your assessment that this is a known/fixed issue.

:shrug: